Fix contract and add tests.
Fix contract.

Update test case.

Refine code.

Update flashloan core code.

+++ contracts/interfaces/IFlashLoanV1Pool.sol
@@ -30,7 +30,7 @@ interface IFlashLoanV1Pool {
         uint amount,
         uint premium
     );
-    event Sync(uint112 reserve);
+    event Sync(uint reserve);
 
     function MINIMUM_LIQUIDITY() external pure returns (uint);
     function factory() external view returns (address);
